- Blood: Floatify `PLAYER::zViewVel`.

+++ source/games/blood/src/player.h
@@ -109,7 +109,7 @@ struct PLAYER
 	int                 bloodlust;  // ---> useless
 	double              zView;
 	double              ozView;
-	int                 zViewVel;
+	double              zViewVel;
 	int                 zWeapon;
 	int                 ozWeapon;
 	int                 zWeaponVel;
